train_metapolyp/requirements.txt
kecam==1.3.17
keras_cv_attention_models==1.3.9
lion_pytorch==0.1.2
numpy==1.24.3
Pillow==10.1.0
Pillow==10.2.0
pytorch_lightning==2.2.0.post0
scikit_learn==1.3.2
torch==2.1.2
torchsummary==1.5.1
torchvision==0.16.2